Add files via upload
Sentiment analysis of mental health is a technique that involves analyzing data obtained from social networking sites such as Twitter to predict depression, anorexia, and other similar mental illnesses among youth12. The importance of detecting depression is that it is the root cause of a plethora of diseases. Early prediction can also mitigate the number of suicides1.

Emotions are certain complicated neurological expressions including three sections: a sentimental occurrence, a psychological and physical reaction, and a sociological or revealing reaction3. Mental state of a person is the perspective of that person and furthermore gives a sign of his/her general nature3.